Refactor to human readable structure
